feat: implement new contact section with responsive form, sidebar cards, and CTA components
